+++ apps/portal-api/src/modules/document/document-download-endpoint.ts
@@ -15,7 +15,7 @@ import { MinIOClient } from '../../thirdparty/minio/client';
 import { logApp } from '../../utils/app-logger.util';
 import { NotFoundError } from '../../utils/error/error.util';
 import { loadOrganizationBy } from '../organization-management/organizations/organizations.domain';
-import { loadUserBy } from '../organization-management/users/users.domain';
+import { loadUserBy } from '../organization-management/users/user-domain/users.domain';
 import { loadServiceDefinitionByServiceInstance } from '../services/service-instance.domain';
 import { telemetryApp } from '../telemetry/telemetry.app';
 import {
