+++ apps/portal-api/src/modules/organization-management/users/user-domain/users.domain.test.ts
@@ -5,22 +5,22 @@ import {
   requestContextAdminUser,
   requestContextSimpleUserSecondOrga,
   TEST_ORGANIZATIONS,
-} from '../../../../tests/tests.const';
+} from '../../../../../tests/tests.const';
 import {
   OrderingMode,
   UserOrdering,
-} from '../../../__generated__/resolvers-types';
-import { requestContext } from '../../../context/request.context';
-import { UserId } from '../../../model/kanel/public/User';
-import { ROLE_ADMIN } from '../../../portal.const';
-import { telemetryApp } from '../../telemetry/telemetry.app';
-import { TelemetrySource } from '../../telemetry/telemetry.const';
+} from '../../../../__generated__/resolvers-types';
+import { requestContext } from '../../../../context/request.context';
+import { UserId } from '../../../../model/kanel/public/User';
+import { ROLE_ADMIN } from '../../../../portal.const';
+import { telemetryApp } from '../../../telemetry/telemetry.app';
+import { TelemetrySource } from '../../../telemetry/telemetry.const';
 import {
   loadUserBy,
   loadUserConnection,
   updateUser,
   updateUserAtLogin,
-} from './users.domain';
+} from '../users.domain';
 
 //Issue with test
 describe('Users domain', () => {
